Tests created. Makefiles updated

+++ src/Hashtable_config.h
@@ -1,6 +1,6 @@
 #ifndef HASHTABLE_CONFIG_H
 #define HASHTABLE_CONFIG_H
 
-typedef double ht_elem_t;
+typedef size_t ht_elem_t;
 
 #endif // HASHTABLE_CONFIG_H
